Complete 1st iteration of the game
The game currently is played with one human
player and a randomly generated master code that
includes duplicates.

It correctly plays the number of guesses selected
at the beginning of the game and displays hints
correctly.

Next iteration would be to let the player choose
the code and the computer guess it.